Fixing issues in forum spellcheck iframe. CSS was not rendering on front end requests.
Commented out a could of console.log()'s that were creating javascript errors, thus killing spellcheck on the front end.

theme switching was not working b/c the spellchecker was being tossed into the session tracker array.
The fix in place is not totally working b/c the entire tracker array is being unset,
but you can now switch themes in the forum without getting redirected to the spellcheck_iframe method.
Marked a @todo and will come back to this.