Fixed a bug (#13975) where a display error occurred in modal dialogues on the file edit page.
file_manager_edit was using 'simple modal' and was a complete mess.  Fixed scoping issues and undefined vars/functions in the file, as well as moved to jquery UI for the modal.  Will be balleting the simplemodal so we are not tempted to use such a thing ever again.  :D