Very quick patch for {body} of forum posts to encode EE tags.
Is an echo of a larger problem of Safe HTML and EE braces being encoded, then unencoded by xss_clean(), thereby not respecting the $convert_curly Typography class property.  Impact is rare since it is not a problem in the Channel module, and requires that a field be set for Safe HTML, and the code parsed by the Template parser, and not have any additional output protection in the module.  Might be limited to the forum module, but a real solution that switches the order of when $convert_curly takes place is probably recommended.