Fixed a bug (#16385) where selected entry categories would be unselected if they were edited from the publish screen.
The selected_cats array would be overwritten each time the form was reloaded, which was each time you clicked an ordering arrow or an edit/delete link. When those reloads happened, the category checkboxes weren't in the DOM anymore to be referenced. So, we need to make sure we're on the category selection screen before trying to reference these checkboxes.