First crack at a deprecated() method for getting data about a method that was deprecated
Added a function to the logger class that can be called from anywhere with the only argument being what version of EE the method was deprecated. From there, it uses debug_backtrace() to actually determine what the deprecated method was and where it was called from. Right now, this method echos a deprecation message at the top of the page and is simply a test of data gathering.

Give it a try, call $this->EE->logger->deprecated(); on your favorite deprecated function today!