Fixed a bug (#17453) where listing entries in an MSM site by category URL title would fail if another site used the same category URL title.
When fixing listing entries by category URL title, we failed to account for the case where multiple category groups may have a category with the same URL title, like with MSM sites. The query above the fix was doing it correctly by checking for the correct cat group ID, so I deleted the old fix and simplified things.