Implementing a solution for the file field querying on every row of the entries loop, which is also something any fieldtype can take advantage of
The problem with fieldtypes that have their own tables is that they likely need to query with each row of the channel entries loop. So I created a solution for fieldtypes to know beforehand exactly what data the entries loop will be parsing, so that the fieldtype can do potentially just one query instead of a query for each row of channel entry data.

I'm going to work on benchmarking to make sure these for loops aren't causing too much trouble, and maybe explore more efficient ways of traversing that data other than nested loops.