Fixing <p>&nbsp;</p> as best we can I think
It basically comes down to we have to have something inside of an element to select it due to bugs in Webkit and IE8. And we have to start off with paragraph tags in the editor or else in Firefox may not be able to get back into a block element after selecting all text and deleting, which is why <p>&nbsp;</p> was introduced in the first place. So we're using a zero-width character in place of the non-breaking space like other rich text editors so the extra space is not apparent.