Fixed a bug (#17627) where submitting a publish form that fails validation may delete data in that entry's file fields.
When a publish form submission fails due to validation, its form elements are populated with the results of the fieldtype's validation process. The file field wasn't returning the data in the same format as it's saved in the database, now it does. This will need to be revised later once we're saving files via the file ID.