Added a hidden configuration variable template_sync_ignore_prefixes for designating text file prefixes that should be ignored when automatically syncing the database to flat files and fixed 'bug' 17403 - Altered the behavior of automatically syncing the database to an existing flat file when   the url is accessed on the frontend to no longer require the template file is writable (see #17403).
Man- I was not comfortable with just removing the is_writable check to address that bug- because in the rare case of that 404 issue that caused the check to go in there?  It has a REALLY bad effect.

So- combined it with this story about skipping ._ prefixed files:
https://www.pivotaltracker.com/story/show/19402315

So now if that guy who had the 404 issue come up comes back- he could just add … 404 to the array.  which actually won't work for him.

Blarg.  We'll just have to tell him to talk to his server admin?