Half-rewrite of the button api. Now setup statically and instantiated per editor with a sane prototype. Read the long description.
Buttons are now added statically to WysiHat instead of directly to the toolbar:
WysiHat.addButton('bold', { .. options .. })

Also notice that the name is now a separate parameter. The editors are then set up with:
$('.rte').wysihat(['bold', 'italic']);

That let's us keep the buttons separate from the setup code.

Buttons now also magically extend a "BlankButton" prototype. This gives them access to:
this.name
this.$editor
this.$element (change to $button?)
this.Selection (selectionUtil)
setOn() / setOff() (hopefully this will help solve the bold bug)

Lastly, HEADINGS ARE STUPID!