Reducing the number of closures. Cleaning up WysiHat.Commands and changing how the editor is set up in regards to those.
No more $.extend($editor, Commands). It was a stupid thing to begin with. While it looks like it gives the editor all of the commands, what it really does is extend the jQuery result set. Meaning you need to have *that one*. It was silly. It's all inside data('wysihat') now.