Fixed a bug (#17906) where the code view of the Rich Text Editor on the front-end may appear too narrow.
We apparently never tested the RTE on the front-end with a class other than .rte, and a lot of styles and some tool code depended on that. So I better-namespaced that class, 'rte' was too generic, and changed it to WysiHat-field, and made sure fields on the front-end got that class on init so the proper styles are applied.