Fixed a bug (#17907) where the Rich Text Editor would pass publish form validation even if field was set to be required. Fixed a bug (#17931) where saving an empty RTE field inserts empty paragraph tags into the database, making template conditionals not work.
It may be nice to find a way to do this in the javascript so that front-end forms aren't inserting empty paragraph tags into the DB too. Not sure the best way to do this at the moment, maybe something in the module JS that binds to its parent form submit to clear out the textarea? Or just not have empty paragraph tags in the textarea when the editor is empty? Pascal? Bueller?