Fixed a problem that crops up if re-running the updater over top of an existing installation.
In testing these changes on a 2.5.2 installation, I'm re-running the updater on it and changing the $config['app_version'] value as I work backwards on the ud_*.php files. Ran into trouble where both 'caption' and 'description' existed after re-running ud_220.php, and the updater borked when it got to ud_240.php. Will likely need to make this part of the check anytime an existing field is renamed.