Prototype of a new feature in Progress lib: update on the step count within each update file.
For now, the Progress library is just displaying this data (the number of steps per update file and which of the steps the update wizard is currently running) on the front-end, but the long view here is that we want to make sure this data is available so that we can improve the update wizard's UX later. For example, this would allow us to display a progress bar with more precision than just big jumps from one update file to the next.