Improved clearing of Tracker cookie when on a Page URL.
This was insane- in 1.x, somehow //$FNS->set_cookie('tracker', serialize($tracker), '0'); was resulting in a disallowed key message.  The key triggering it was, for example:
s:11:"/home/test/"

Which yes, should indeed trigger that warning.  If I deleted the tracker cookie and commented out this line, then there was no further problem.

This line basically just deletes the cookie.  WHY do we delete if for a PAGE?  Makes no sense to me.  But if we want to delete it, new code is better- and does not cause that error.

But the error?  Still puzzles me.  As does why we want to delete the tracker array when we hit a Page page.