Reverting fix for bug #16832 to fix bug #18263.
This was a messy one, but the original bug fix was just to please Paul by removing some PHP 4 code. But removing that code led to strange unreproducible PHP notices for some folks. Should not have been happening logically, but could not figure it out. So, reverted for now.

This reverts commit 743df6b094021de3662b9dfb109c777245d17988.

Conflicts:
	changelogs/ee_changelog.txt