Fixed a bug (#16985) where ExpressionEngine tags were not encoded in some category fields.
Some folks could actually have been using the fact tags would be parsed here.  However, it honestly seems both inconsistent and a possible security issue to me.  You shouldn't be able to put a query tag in as a category name and have it parsed.

Worst case scenario, if someone was using it, we have add ons that will allow the tag to be parsed.