Finishing deprecation of Localize::fetch_date_params, should never need it again
The old way of parsing date formats was to fetch an array of all the date formatting characters, then replace them one by one  in the template; all of this was up to the module to do, quite silly, now you can jus pass in the the entire format string to the Localize class and it will take care of it.

The whole handling of parsing out the format string from the date tag is a another story. Many modules are handling it themselves when they shouldn't have to, the same boiler plate code and regex is copy and pasted everywhere to parse out the value of a format= parameter. Will have to revisit.