Added channel_search_modify_query hook to mod.search.php
This is a little self serving, but I can't use the native search functionality with Publisher without a hook. It lets the developer modify the query string used in the search function. If end_script is true, then it returns early and saves the custom query to the db table, not the "clean" query at the bottom of the method. I've tested this and it appears to be working fine. Example usage: https://gist.github.com/litzinger/e18be6b3b30a6f9d6042