Removing the ability to change session ttls. Adding expire_session_on_browser_close config.
A lot of people change these configs out of frustration and end up
compromising the security of their site. There are two reasons to
change session ttls:
The first is to achieve longer sessions, where it is safer to use
remember me.
The second is to achieve sessions that end on browser close. That’s
what the new config is for.