Making the di a dependency for the model again so it can be flipped out [cont.]
It is only at the top level (model builder). This keeps it exchangeable
while avoiding the testing cost of the law of demeter violations.
Always be given dependencies. If you need to ask for a dependency, then
the layer above yours should have a factory that asks for it. See
model::newGateway change in this commit for an example. Note: mixing
factories like that is ugly - ideally you can get your factory passed
in from the di and just incur the cost when it’s not used.