Changing greater_than rules on crop validation to is_natural_no_zero
greater_than doesn't have a lang key and I'm unsure how to generate one that tells the user what the field needs to be greater than. Might need to be a lang key we define on a case-by-case basis, but can just use is_natural_no_zero if we simply want greater than zero.