Per Kevin, checking for $this->Author to prevent a bug in layouts....
To quote: t looks like in the Layouts controller we're creating a dummy ChannelEntry object to get a few things out of it, but it doesn't have an author, so it's throwing this error. May just need to put a check in that populateAuthors authors method to make sure there's an author.

I did get it to throw the error- the tweak eliminates it an all seems fine on publish still.