Proposed fix for meta refresh Windows 10/Edge
See https://github.com/EllisLab/ExpressionEngine/issues/1366#issuecomment-268287850

This just sticks a js refresh in there that fires just before the meta refresh.

I couldn't for the life of me spot why it fails and it's got to be rare, as it came up only a very few times googling.