:bug: don't save FieldModel unless we're creating or editing a field
Inside `save()` this was being called from its association to the
Channel model, which was in turn called (to update stats) when saving a
ChannelEntry.

The result is that `post_save_settings()` was being called in any
fieldtype that has that method. In the case of Grid, the presence of
`$_POST[‘grid’]` from Channel Form if you had a field _named_ `grid`
was causing all columns to be dropped. :ohmy:

That’s a mouthful, but is the simplest way I can explain it. This
change should ensure that that method only fires when a field itself is
updated or created.