Complete (?) fix for Gzip and PHP errors
ref: 6af6171

Essentially, @ suppressed errors still load our Exceptions class and
are potentially logged, so we need to account for actual PHP error
display instead of just the existence of the class.

Template Routes use @preg_match() dozens of times on invalid strings,
so just using the template routes feature ensures that the Exceptions
class is loaded and tons of @ errors are generated.