Hook up user notifications
I really don’t like this code. I want a new shared service for
notifications but a minor release is not right. So this is just
extracted from the comment module so the same code can be used
regardless of when the notifications are sent and from where. Some
minor refactoring so that email template parsing is these same for
admin and user notifications, with no feature changes, but ugh are
these data structures from the legacy comment model gross. :bandaid: