Parse all category variables in the same place, fix formatting
Cat custom fields were not respecting their formatting as they were
being replaced with a blunt str_replace() in the entries Parser. This
stops that, lets all category variables be parsed with the same logic.
Includes a change to the FieldModel with unknown implications. We call
the setup_handler and overwrite the field_types array of the channel
fields API so many times and from so many different places it makes my
head spin. But calling the setup handler again in this case was causing
a nulled copy of the field type to be used in apply() rather than the
one we bothered setting up as the $fieldtype in this method. Since I
have no idea how many different places are using this “TEMPORARY,
VOLATILE, DO NOT USE” method, I have no idea what I just broke. Yey.