Added timestamp variables to the default date fields in the channel form for better date format compatibility in the date picker.
dd/mm/yyyy  caused the datepicker to error and not show dates for entry_date, expiration_date and comment_expiration date.  Worked for custom date fields.

Adding the data-timestamp= to the field fixed it and was consistent with the custom fields, but we needed a timestamp variable to populate it.