Fix fatal error when upgrading from < 2.11
The method update_grid_search() was never added to v3, it was a late patch to v2. This update bug was introduced in v4 here, where I wanted to make sure updates were all in parity: aebc091

Will add a new fix for v4 that is modern and works with the new table structure.