Fix fatal error when upgrading from < 2.11
The method update_grid_search() was never added to v3, it was a late patch to v2. This update bug was introduced in v3 here, where I wanted to make sure updates were all in parity: 97377b3

It isn't a data loss bug, so #wontfix in v3, just keep it from preventing an upgrade.