Loosen gate for returning validation message in Grid
This makes it so if you trigger a change event on an input that isn’t named exactly what the base name should be, we’ll still return your validation message. This mirrors behavior for non-namespaced fields where we will strip any [] modifications to send validation to the base field and return its message.

https://expressionengine.com/forums/topic/251210/grid-field-validation-with-ansel