Made {path:image_url} custom theme aware css_file_path = ‘’
Member was hard coding the path to images so it always pointed to themes/ee.  I switched to use the service and it should pick up custom themes now.

The css_file_path in both members and themes seems to point to non-existent files.  The css is in templates, not assets.

I could make it point to the folder instead of blank, but it’s just wrong the way it was.