Make sure response_date is always DateTime
Small inconsistency between timestamp columns retreived from the database and those that have just been saved to the database.

On save(), timestamp columns are int. If we try to operate on them after saving in the same request, we get fatal PHP errors for trying to use an int as an object.