Revert "If the `channel=` param is present use it’s channel_ids array, don’t obliterate it."
This reverts commit 793fffd32aa7a39474a37292927a5010da004a0e.

The errors for which this commit was intended to fix were fixed in other
ways later. This commit introduced some performance problems in Channel
Entries tags that referenced mutiple channels, causing the Channel
module to query for field infomation for channels that were not returned
by the tag, and also the fields from other sites if another site had the
same channel name.