Do not populate all field defaults on entry data set
Hello, developer of the future! You've likely arrived here via git-bisect, and
for that, I'm sorry. My goal here was to make ChannelEntry models more
lightweight to use when not needing to populate a publish form, and a great way
to do that seemed to be to avoid populating all the default channel fields when
data is set via setData(). After doing archaeology on that line, I think the
problems it was trying to solve are no longer present. But, we'll see. And as
for the call on save(), the call seemed redundant when initField() also calls
the method.