Account for PHP's replacement of dots in GET variable names in Template Manager
Since the template manager loads template groups via its group name in a
URL segment, if you had a group name containing a period, it would get
turned into an underscore and we couldn't find it. So in that case, we'll
replace underscores with dots and try again. This doesn't account for the
case where a group name contains both a dot and an underscore, as we can't
very well know which underscores need replacing. So, that's a bug for
another day.

Fixes #166