Fixing object soft reserved keyword
'object' is a soft reserved keyword as of PHP version 7.0 and a
reserved keyword as of PHP version 7.2 and should not be used to name
a class, interface or trait or as part of a namespace (T_NAMESPACE).

Chose to make change to namespace, since this is the only place this
is utilized.