Add pass-through facility for HTML forms
Modify the form_declaration function in Functions.php.

Updated version fixes a couple of minor errors in original code, and adds a new attribute pass-through function.  When called the function inspects the current tagparams and if any of the parameters specified match valid HTML attributes for use in a form it creates these in the opening <FORM> tag including the parameter value as an attribute value if one is provided.  If no parameter value provided the attribute is included as a token (e.g. 'novalidate').

If one of the parameters in the tab matches one that is possibly also assigned within the calling function the value assigned within the calling function is given priority except in the case of the parameter "class=" - in this case the value of the additional class paramter is appended to the value generated by the calling function.