Fixed footer link permissions and tweaked language
The link to support didn’t have a permission check on it- before, that’s where ‘submit a ticket’ went.  I added back the permission check and I also tweaked the description in the permission settings to say ‘Get Support’ as it’s a bit more clear what you’re doing.  And it fits with the language used in the footer.