Remove CSS styling which sets status to uppercase
The status column in the list of entries shows the status items in
uppercase. This is caused by a text-transform property in the CSS for
the "status-tag" class.

The status column is useful as a reference to be used in channel entry
tags, but the CSS obscures the actual casing being used.

Change the text-transform property from uppercase to "none." This change
affects status items being used elsewhere, creating the same issue in
each of these instances.