Show any errors for resending activation emails
When resending activation emails, errors should be returned when the
email fails to send. This behavior should match the functionality of
sending emails using the communicate utility. If the email fails to send,
the received error gets displayed. This gives the user an indicator that
the email failed to send along with information which can be useful for
troubleshooting. Currently there is no indication that an email failed
to send when resending these activation emails.

Add a conditional to the email send request similar to that which is used
for the communicate functionality.