Restrict textarea resize to vertical
The current textarea size is set to a min and max of 100%. Because the
textarea is always set to the max width of the parent element, then it
shouldn't need to be resized horizontally. Resizing is an automatic
feature in some browsers and may be overridden.

In some cases, resizing the textarea outside the parent element creates
issues with other child elements which share the same parent. As an
example, the issue this change is meant to address is that a resized
text area is pushing "delete" controls for text inputs off the screen.
The user would then have to refresh the page to get those controls
back.

Note that the user is unable to make the textarea smaller because of
other styling rules. With both max and min width set to 100%, the user
can only make the textarea larger. Though this would be an alternative
for fixing this issue, it's reasonable to assume that we should be
constraining the size to the max width of the parent.