Add ability to use class/method to ACT requests
This PR modifies existing code to add the ability for developers to access ACT requests using the Class/Method reference instead of making calls to fetch the action id. It also moves the conditional to check if the module is installed outside the main conditional (the current ordering would only check if the ACT was not numeric, which would ignore many calls.