More detailed instructions for running Cypress (#1671)
* More detailed instructions for running Cypress

Everything in this file "worked" for me the first time, except when I went to actually run the tests locally it kept failing on the db:seed task. It would be helpful if this file was more detailed so someone with no Cypress experience could setup and run the tests locally without error. 

Does it require a minimum version of Node/NPM? If so it should be mentioned in the README.

Does the user need to have local instance of MySQL server running on their machine? If so this should be specifically mentioned and not assumed.

Also fixed the path to the config.php file.

* extra wait on tests

Co-authored-by: Yuri Salimovskiy <yuri@packettide.com>