If an add-on url is wrong it was throwing a php error
Better to throw a 404 than show an error (which was confusing).

We were getting a 'Call to a member function checkCachedLicenseResponse() on null' in Controller/Addons/Addons.php line 745 when we had a malformed add-on url.  Tossed the other 404s in as well since they just made sense to me, but probably not needed.