Resolved #1851 where some custom fields could be excluded from search
The solutions here is to only fetch searchable fields for the channels that are searched. This should significantly reduce the size of query for most of the sites.
The bigger sites (50+ searchable fields on channel) would still have the issue - we will write warning to developer log, because searching 50+ fields without index is really bad for performance anyway