Had wrong name for a low reorder hook
Low reorder has an extension called 'low_search_post_search' that can be called to alter pro search ordering.

Looks like when we updated from low_search to pro_search we did a find/replace and replaced the hook call when we shouldn't have.  And since it's the low reorder hook we're calling, we need to call the correct hook.