Update pi.rss_parser.php
The docs say that {if no_results} works, but it didn't. If there were no items, it still showed the feed meta data and just nothing for the missing items.  If the tag was present, it and the contents were removed.

I'd personally like a way to show the feed meta data and a no results type tag for the items, but that's not consistent with what docs say or what user needed and given I haven't had any request for that, I'm being pedantic.

Hence- this simple 'Make no results work like normal' pr.