Download and install update for add-on
This is a follow-up for #5099

Currently prototype. as it lacks endpoint that needs to be built in EE store

If the add-on has update available in License Store (license returns as update_available) it will show the banner with the link. Clicking on the link will have the add-on downloaded, unpacked and updated.

Note: the "update available" is currently only shown for add-ons that have Settings page. We need to expand that.
The server side of downloader is yet to be implemented on EE Store side. It needs to serve the add-on zip with the file name matching add-on name.
We also want hash file to verify the intergrity, built using same algorythm as EE update hash file.